+++ gateway/handlers/http/login.go
@@ -15,9 +15,10 @@ import (
 )
 
 func login(w http.ResponseWriter, r *http.Request) {
+	panic("")
 	req := &dto.LoginRequest{}
 	ctx := r.Context()
-	translate := ctx.Value("translate").(translator.TranslatorFunc)
+	translate := ctx.Value(g.TranslateContext).(translator.TranslatorFunc)
 	utils.ParseBody(r.Body, req)
 	utils.ValidateBody(req, dto.LoginValidator, translate)