+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- 🐛 fix: bug fixed for context key type(`translate` key)
 - 🎉 feat: added auth microservice with Login, Register and UniquePhoneNumber methods
 - 🎉 feat: auth microservice added
 - 🎉 feat: fixed gateway